# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/schedules": {
      "post": {
        "operationId": "SchedulesController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Schedule"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Schedules"]
      },
      "get": {
        "operationId": "SchedulesController_findAll_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "staffId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InfinityPaginationScheduleResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Schedules"]
      }
    },
    "/api/v1/schedules/by-day": {
      "get": {
        "operationId": "SchedulesController_findByDay_v1",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Schedule"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Schedules"]
      }
    },
    "/api/v1/schedules/{id}": {
      "get": {
        "operationId": "SchedulesController_findById_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Schedule"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Schedules"]
      },
      "patch": {
        "operationId": "SchedulesController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScheduleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Schedule"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Schedules"]
      },
      "delete": {
        "operationId": "SchedulesController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Schedules"]
      }
    },
    "/api/v1/schedules/staff/{staffId}": {
      "get": {
        "operationId": "SchedulesController_findByStaff_v1",
        "parameters": [
          {
            "name": "staffId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Schedule"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Schedules"]
      }
    },
...
    "/api/v1/appointments": {
      "post": {
        "operationId": "AppointmentsController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Appointment"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Appointments"]
      },
      "get": {
        "operationId": "AppointmentsController_findAll_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "startTime",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "endTime",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InfinityPaginationAppointmentResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Appointments"]
      }
    },
    "/api/v1/appointments/countCustomer": {
      "get": {
        "operationId": "AppointmentsController_countByCustomer_v1",
        "parameters": [
          {
            "name": "startTime",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "endTime",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentSatisticDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Appointments"]
      }
    },
    "/api/v1/appointments/count": {
      "get": {
        "operationId": "AppointmentsController_count_v1",
        "parameters": [
          {
            "name": "startTime",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "endTime",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentSatisticDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Appointments"]
      }
    },
    "/api/v1/appointments/{id}": {
      "get": {
        "operationId": "AppointmentsController_findById_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Appointment"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Appointments"]
      },
      "patch": {
        "operationId": "AppointmentsController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Appointment"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Appointments"]
      },
      "delete": {
        "operationId": "AppointmentsController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Appointments"]
      }
    },
...
    "/api/v1/receptions/appointment/{appointmentId}": {
      "get": {
        "operationId": "ReceptionsController_findByAppointment_v1",
        "parameters": [
          {
            "name": "appointmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Receptions"]
      }
    },
...
  },
  "info": {
    "title": "API",
    "description": "API docs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "FileType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cbcfa8b8-3a25-4adb-a9c6-e325f0d0f3ae"
          },
          "path": {
            "type": "string",
            "example": "https://example.com/path/to/file.jpg"
          }
        },
        "required": ["id", "path"]
      },
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string",
            "example": "admin"
          }
        },
        "required": ["id", "name"]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "zaloId": {
            "type": "string",
            "example": "1234567890"
          },
          "phoneNumber": {
            "type": "string",
            "example": "1234567890"
          },
          "userName": {
            "type": "string",
            "example": "John"
          },
          "avatar": {
            "example": "avatar image link",
            "allOf": [
              {
                "$ref": "#/components/schemas/FileType"
              }
            ]
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "zaloId",
          "phoneNumber",
          "userName",
          "avatar",
          "role",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
...
      "StatusEnum": {
        "type": "number",
        "enum": [1, 2, 3, 4, 5, 6, 7, 8],
        "description": "Status of the appointment (1 = pending, 2 = inprogress, 3 = done)"
      },
      "CustomerRecord": {
        "type": "object",
        "properties": {
          "active": {
            "type": "boolean",
            "nullable": false
          },
          "BHYTNumber": {
            "type": "string",
            "nullable": true
          },
          "CCCDNumber": {
            "type": "string",
            "nullable": true
          },
          "DOB": {
            "format": "date-time",
            "type": "string",
            "nullable": false
          },
          "sex": {
            "type": "string",
            "nullable": false
          },
          "fullName": {
            "type": "string",
            "nullable": false
          },
          "user": {
            "nullable": false,
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "active",
          "BHYTNumber",
          "CCCDNumber",
          "DOB",
          "sex",
          "fullName",
          "user",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
      "Appointment": {
        "type": "object",
        "properties": {
          "specificTime": {
            "format": "date-time",
            "type": "string",
            "nullable": false
          },
          "active": {
            "type": "boolean",
            "nullable": false
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "description": "Status of the appointment (1 = pending, 2 = inprogress, 3 = done)",
            "example": 1,
            "allOf": [
              {
                "$ref": "#/components/schemas/StatusEnum"
              }
            ]
          },
          "customerRecord": {
            "nullable": false,
            "allOf": [
              {
                "$ref": "#/components/schemas/CustomerRecord"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "specificTime",
          "active",
          "note",
          "status",
          "customerRecord",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "StaffDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": ["id"]
      },
...
      "CreateScheduleDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string"
          },
          "endTime": {
            "format": "date-time",
            "type": "string"
          },
          "startTime": {
            "format": "date-time",
            "type": "string"
          },
          "staff": {
            "$ref": "#/components/schemas/StaffDto"
          }
        },
        "required": ["endTime", "startTime", "staff"]
      },
      "Staff": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "nullable": true
          },
          "specialization": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "user": {
            "example": "userId",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "note",
          "specialization",
          "id",
          "user",
          "createdAt",
          "updatedAt"
        ]
      },
      "Schedule": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "nullable": true
          },
          "active": {
            "type": "boolean",
            "nullable": false
          },
          "endTime": {
            "format": "date-time",
            "type": "string",
            "nullable": false
          },
          "startTime": {
            "format": "date-time",
            "type": "string",
            "nullable": false
          },
          "staff": {
            "nullable": false,
            "allOf": [
              {
                "$ref": "#/components/schemas/Staff"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "note",
          "active",
          "endTime",
          "startTime",
          "staff",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
      "InfinityPaginationScheduleResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Schedule"
            }
          },
          "hasNextPage": {
            "type": "boolean",
            "example": true
          }
        },
        "required": ["data", "hasNextPage"]
      },
      "UpdateScheduleDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string"
          },
          "endTime": {
            "format": "date-time",
            "type": "string"
          },
          "startTime": {
            "format": "date-time",
            "type": "string"
          },
          "staff": {
            "$ref": "#/components/schemas/StaffDto"
          }
        }
      },
...
      "ServiceScheduleDto": {
        "type": "object",
        "properties": {
          "serviceId": {
            "type": "string",
            "example": "serviceId"
          },
          "scheduleId": {
            "type": "string",
            "example": "scheduleId"
          },
          "staffId": {
            "type": "string",
            "example": "staffId"
          }
        },
        "required": ["serviceId", "scheduleId", "staffId"]
      },
      "CustomerRecordDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": ["id"]
      },
      "CreateAppointmentDto": {
        "type": "object",
        "properties": {
          "specificTime": {
            "format": "date-time",
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "serviceAndScheduleIds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServiceScheduleDto"
            }
          },
          "customerRecord": {
            "$ref": "#/components/schemas/CustomerRecordDto"
          }
        },
        "required": ["specificTime", "serviceAndScheduleIds", "customerRecord"]
      },
      "InfinityPaginationAppointmentResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Appointment"
            }
          },
          "hasNextPage": {
            "type": "boolean",
            "example": true
          }
        },
        "required": ["data", "hasNextPage"]
      },
      "AppointmentSatisticDto": {
        "type": "object",
        "properties": {
          "count": {
            "type": "number"
          }
        },
        "required": ["count"]
      },
      "UpdateAppointmentDto": {
        "type": "object",
        "properties": {
          "specificTime": {
            "format": "date-time",
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "serviceAndScheduleIds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServiceScheduleDto"
            }
          },
          "customerRecord": {
            "$ref": "#/components/schemas/CustomerRecordDto"
          },
          "status": {
            "description": "Status of the appointment (1 = pending, 2 = inprogress, 3 = done)",
            "example": 1,
            "allOf": [
              {
                "$ref": "#/components/schemas/StatusEnum"
              }
            ]
          }
        },
        "required": ["status"]
      },
...
      "Reception": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "description": "Status of the appointment (1 = notCheckin, 2 = checkin)",
            "example": 1,
            "allOf": [
              {
                "$ref": "#/components/schemas/StatusEnum"
              }
            ]
          },
          "checkinTime": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "Appointment": {
            "nullable": false,
            "allOf": [
              {
                "$ref": "#/components/schemas/Appointment"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "note",
          "status",
          "checkinTime",
          "Appointment",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}